data analysis
this data is taken from kaggle. and i have used this data to achieve the answer of following question's answer like, who is the best player? which team is most consistent since 2008 to 2019? which teams are top 3 performer? and the dependency of match on toss winner and their decision.


i have used tableau to perform cleaning and excel for visualization. i have created an interactive dashboard for the same objectives.